+++ lib/Data/PrimitiveArray/Class.hs
@@ -30,67 +30,83 @@ import           Data.PrimitiveArray.Index.Class
 
 data family MutArr (m :: * -> *) (arr :: *) :: *
 
+-- | Associate a fill structure with each type of array (dense, sparse, ...).
+--
+-- Example: @type instance FillStruc (Sparse w v sh e) = (w sh)@ associates the type @(w sh)@, which
+-- is of the same type as the underlying @w@ structure holding index information for a sparse array.
 
--- | The core set of operations for monadic arrays.
-
-class (Index sh) => MPrimArrayOps arr sh elm where
-
-  -- | Return the bounds of the array. All bounds are inclusive, as in
-  -- @[lb..ub]@
-
-  upperBoundM :: MutArr m (arr sh elm) -> LimitType sh
-
-  -- | Given lower and upper bounds and a list of /all/ elements, produce a
-  -- mutable array.
-
-  fromListM :: PrimMonad m => LimitType sh -> [elm] -> m (MutArr m (arr sh elm))
-
-  -- | Creates a new array with the given bounds with each element within the
-  -- array being in an undefined state.
-
-  newM :: PrimMonad m => LimitType sh -> m (MutArr m (arr sh elm))
-
-  -- | Creates a new array with all elements being equal to 'elm'.
-
-  newWithM :: PrimMonad m => LimitType sh -> elm -> m (MutArr m (arr sh elm))
-
-  -- | Reads a single element in the array.
-
-  readM :: PrimMonad m => MutArr m (arr sh elm) -> sh -> m elm
-
-  -- | Writes a single element in the array.
-
-  writeM :: PrimMonad m => MutArr m (arr sh elm) -> sh -> elm -> m ()
+type family FillStruc arr :: *
 
 
 
--- | The core set of functions on immutable arrays.
+-- | The core set of operations for pure and monadic arrays.
 
 class (Index sh) => PrimArrayOps arr sh elm where
 
-  -- | Returns the bounds of an immutable array, again inclusive bounds: @ [lb..ub] @.
+  -- ** Pure operations
 
+  -- | Returns the bounds of an immutable array, again inclusive bounds: @ [lb..ub] @.
   upperBound :: arr sh elm -> LimitType sh
 
-  -- | Freezes a mutable array an returns its immutable version. This operation
-  -- is /O(1)/ and both arrays share the same memory. Do not use the mutable
-  -- array afterwards.
-
-  unsafeFreeze :: PrimMonad m => MutArr m (arr sh elm) -> m (arr sh elm)
-
-  -- | Thaw an immutable array into a mutable one. Both versions share
-  -- memory.
-
-  unsafeThaw :: PrimMonad m => arr sh elm -> m (MutArr m (arr sh elm))
-
-  -- | Extract a single element from the array. Generally unsafe as not
-  -- bounds-checking is performed.
-
+  -- | Extract a single element from the array. Generally unsafe as not bounds-checking is
+  -- performed.
   unsafeIndex :: arr sh elm -> sh -> elm
 
-  -- | Savely transform the shape space of a table.
+  -- | Index into immutable array, but safe in case @sh@ is not part of the array.
+  safeIndex :: arr sh elm -> sh -> Maybe elm
 
+  -- | Savely transform the shape space of a table.
   transformShape :: (Index sh') => (LimitType sh -> LimitType sh') -> arr sh elm -> arr sh' elm
+
+  -- ** Monadic operations
+
+  -- | Return the bounds of the array. All bounds are inclusive, as in @[lb..ub]@. Technically not
+  -- monadic, but rather working on a monadic array.
+  upperBoundM :: MutArr m (arr sh elm) -> LimitType sh
+
+  -- | Given lower and upper bounds and a list of /all/ elements, produce a mutable array.
+  fromListM :: PrimMonad m => LimitType sh -> [elm] -> m (MutArr m (arr sh elm))
+
+  -- | Creates a new array with the given bounds with each element within the array being in an
+  -- undefined state.
+  newM :: PrimMonad m => LimitType sh -> m (MutArr m (arr sh elm))
+
+  -- | Variant of 'newM' that requires a fill structure. Mostly for special / sparse structures
+  -- (hence the @S@, also to be interpreted as "safe", since these functions won't fail with sparse
+  -- structures).
+  newSM :: (Monad m, PrimMonad m) => LimitType sh -> FillStruc (arr sh elm) -> m (MutArr m (arr sh elm))
+
+  -- | Creates a new array with all elements being equal to 'elm'.
+  newWithM :: PrimMonad m => LimitType sh -> elm -> m (MutArr m (arr sh elm))
+
+  -- | Variant of 'newWithM'
+  newWithSM :: (Monad m, PrimMonad m) => LimitType sh -> FillStruc (arr sh elm) -> e -> m (MutArr m (arr sh elm))
+
+  -- | Reads a single element in the array.
+  readM :: PrimMonad m => MutArr m (arr sh elm) -> sh -> m elm
+
+  -- | Read from the mutable array, but return @Nothing@ in case @sh@ does not exist. This will
+  -- allow streaming DP combinators to "jump" over missing elements.
+  --
+  -- Should be used with @Stream.Monadic.mapMaybe@ to get efficient code.
+  vreadM :: (Monad m, PrimMonad m) => MutArr m (arr sh elm) -> sh -> m (Maybe elm)
+
+  -- | Writes a single element in the array.
+  writeM :: PrimMonad m => MutArr m (arr sh elm) -> sh -> elm -> m ()
+
+  -- | Write into the mutable array, but if the index @sh@ does not exist, silently continue.
+  writeSM :: (Monad m, PrimMonad m) => MutArr m (arr sh elm) -> sh -> elm -> m ()
+
+  -- | Freezes a mutable array an returns its immutable version. This operation is /O(1)/ and both
+  -- arrays share the same memory. Do not use the mutable array afterwards.
+  unsafeFreezeM :: PrimMonad m => MutArr m (arr sh elm) -> m (arr sh elm)
+
+  -- | Thaw an immutable array into a mutable one. Both versions share memory.
+  unsafeThawM :: PrimMonad m => arr sh elm -> m (MutArr m (arr sh elm))
+
+
+
+
 
 class (Index sh) => PrimArrayMap arr sh e e' where
 
@@ -109,36 +125,30 @@ instance Show PAErrors where
 
 
 
--- | Infix index operator. Performs minimal bounds-checking using assert in
--- non-optimized code.
+-- | Infix index operator. Performs minimal bounds-checking using assert in non-optimized code.
 --
 -- @(!)@ is rewritten from phase @[1]@ onwards into an optimized form. Before, it uses a very slow
 -- form, that does bounds checking.
---
--- TODO use @(!?)@ and fail on @Nothing@. Still rewrites in optimized code!
 
 --(!) :: (HasCallStack, PrimArrayOps arr sh elm) => arr sh elm -> sh -> elm
 (!) :: (PrimArrayOps arr sh elm) => arr sh elm -> sh -> elm
 {-# Inline [1] (!) #-}
 {-# Rules "unsafeIndex" [2] (!) = unsafeIndex #-}
-(!) = \arr idx ->
-  let a = inBounds (upperBound arr) idx
-      i = unsafeIndex arr idx
-  in if a then i else traceShow (showBound (upperBound arr), showIndex idx) $ assert a i
+(!) = \arr idx -> case safeIndex arr idx of
+          Nothing -> error $ show (showBound (upperBound arr), showIndex idx)
+          Just v  -> v
 
 
 
 -- | Return value at an index that might not exist.
---
--- TODO implement @maybeIndex@ in the @Ops@ class, helps with sparse arrays too.
 
 (!?) :: PrimArrayOps arr sh elm => arr sh elm -> sh -> Maybe elm
 {-# Inline (!?) #-}
-(!?) arr idx = if inBounds (upperBound arr) idx then Just (arr ! idx) else Nothing
+(!?) = safeIndex
 
 -- | Returns true if the index is valid for the array.
 
-inBoundsM :: (Monad m, MPrimArrayOps arr sh elm) => MutArr m (arr sh elm) -> sh -> Bool
+inBoundsM :: (Monad m, PrimArrayOps arr sh elm) => MutArr m (arr sh elm) -> sh -> Bool
 inBoundsM marr idx = inBounds (upperBoundM marr) idx
 {-# INLINE inBoundsM #-}
 
@@ -160,7 +170,7 @@ inBoundsM marr idx = inBounds (upperBoundM marr) idx
 -- default element, and a list of associations.
 
 fromAssocsM
-  :: (PrimMonad m, MPrimArrayOps arr sh elm)
+  :: (PrimMonad m, PrimArrayOps arr sh elm)
   => LimitType sh -> elm -> [(sh,elm)] -> m (MutArr m (arr sh elm))
 fromAssocsM ub def xs = do
   ma <- newWithM ub def
@@ -173,13 +183,13 @@ fromAssocsM ub def xs = do
 -- | Initialize an immutable array but stay within the primitive monad @m@.
 
 newWithPA
-  ∷ (PrimMonad m, MPrimArrayOps arr sh elm, PrimArrayOps arr sh elm)
-  ⇒ LimitType sh
-  → elm
-  → m (arr sh elm)
+  :: (PrimMonad m, PrimArrayOps arr sh elm)
+  => LimitType sh
+  -> elm
+  -> m (arr sh elm)
 newWithPA ub def = do
   ma ← newWithM ub def
-  unsafeFreeze ma
+  unsafeFreezeM ma
 {-# Inlinable newWithPA #-}
 
 -- | Safely prepare a primitive array.
@@ -188,15 +198,15 @@ newWithPA ub def = do
 -- should see this once the test with NeedlemanWunsch is under way).
 
 safeNewWithPA
-  ∷ forall m arr sh elm 
-  . (PrimMonad m, MonadError PAErrors m, MPrimArrayOps arr sh elm, PrimArrayOps arr sh elm)
-  ⇒ LimitType sh
-  → elm
-  → m (arr sh elm)
+  :: forall m arr sh elm
+  . (PrimMonad m, MonadError PAErrors m, PrimArrayOps arr sh elm)
+  => LimitType sh
+  -> elm
+  -> m (arr sh elm)
 safeNewWithPA ub def = do
   case runExcept $ sizeIsValid maxBound [totalSize ub] of
-    Left  (SizeError _) → throwError PAEUpperBound
-    Right (CellSize  _) → newWithPA ub def
+    Left  (SizeError _) -> throwError PAEUpperBound
+    Right (CellSize  _) -> newWithPA ub def
 {-# Inlinable safeNewWithPA #-}
 
 
@@ -208,22 +218,22 @@ assocs arr = unId . SM.toList $ assocsS arr
 
 -- | Return all associations from an array.
 
-assocsS ∷ forall m arr sh elm . (Monad m, IndexStream sh, PrimArrayOps arr sh elm) ⇒ arr sh elm → SM.Stream m (sh,elm)
+assocsS :: forall m arr sh elm . (Monad m, IndexStream sh, PrimArrayOps arr sh elm) => arr sh elm -> SM.Stream m (sh,elm)
 assocsS arr = SM.map (\k -> (k,unsafeIndex arr k)) $ streamUp zeroBound' (upperBound arr)
 {-# INLINE assocsS #-}
 
 -- | Creates an immutable array from lower and upper bounds and a complete list
 -- of elements.
 
-fromList :: (PrimArrayOps arr sh elm, MPrimArrayOps arr sh elm) => LimitType sh -> [elm] -> arr sh elm
-fromList ub xs = runST $ fromListM ub xs >>= unsafeFreeze
+fromList :: (PrimArrayOps arr sh elm) => LimitType sh -> [elm] -> arr sh elm
+fromList ub xs = runST $ fromListM ub xs >>= unsafeFreezeM
 {-# INLINE fromList #-}
 
 -- | Creates an immutable array from lower and upper bounds, a default element,
 -- and a list of associations.
 
-fromAssocs :: (PrimArrayOps arr sh elm, MPrimArrayOps arr sh elm) => LimitType sh -> elm -> [(sh,elm)] -> arr sh elm
-fromAssocs ub def xs = runST $ fromAssocsM ub def xs >>= unsafeFreeze
+fromAssocs :: (PrimArrayOps arr sh elm) => LimitType sh -> elm -> [(sh,elm)] -> arr sh elm
+fromAssocs ub def xs = runST $ fromAssocsM ub def xs >>= unsafeFreezeM
 {-# INLINE fromAssocs #-}
 
 -- -- | Determines if an index is valid for a given immutable array.
@@ -255,6 +265,6 @@ instance Applicative m => FreezeTables m Z where
 
 instance (Functor m, Applicative m, Monad m, PrimMonad m, FreezeTables m ts, PrimArrayOps arr sh elm) => FreezeTables m (ts:.MutArr m (arr sh elm)) where
     type Frozen (ts:.MutArr m (arr sh elm)) = Frozen ts :. arr sh elm
-    freezeTables (ts:.t) = (:.) <$> freezeTables ts <*> unsafeFreeze t
+    freezeTables (ts:.t) = (:.) <$> freezeTables ts <*> unsafeFreezeM t
     {-# INLINE freezeTables #-}